\name{AdvRknapsack-package}
\alias{AdvRknapsack-package}
\alias{AdvRknapsack}
\docType{package}
\title{
\packageTitle{AdvRknapsack}
}
\description{
\packageDescription{AdvRknapsack}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{AdvRknapsack}

Maintainer: \packageMaintainer{AdvRknapsack}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = AdvRknapsack)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{AdvRknapsack}}
%\section{Documentation Index}{\packageIndices{AdvRknapsack}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
