% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solution to knapsack problem using Brute-Force Algorithm}
\usage{
brute_force_knapsack(x, W, ...)
}
\arguments{
\item{x}{Data.frame Contains 2 columns "w" and "v" containing the weight of sacks and value of sacks respectively.}

\item{W}{Numeric. Specifies the capacity of knapsack.}

\item{...}{Optional logical parameter parallel that defaults to FALSE. When set to true performs parallelised execution of brute force algorithm.}
}
\value{
Named List
}
\description{
The function implements solution to knapsack problem using Brute-Force Algorithm [wiki](https://en.wikipedia.org/wiki/Knapsack).
}
\examples{
#Prepare Data:
suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
knapsack_objects<-data.frame(w=sample(1:4000,size=2000,replace=TRUE),v=runif(n=2000,0,10000))

ptm <- proc.time()
a<-brute_force_knapsack(x = knapsack_objects[1:16,], W = 2000)
print(proc.time() - ptm) # 0.404 milliseconds
print(a)

ptm <- proc.time()
library(parallel)
b<-brute_force_knapsack(x = knapsack_objects[1:16,], W = 2000,parallel=TRUE)
print(proc.time() - ptm) # 0.104 milliseconds
print(b)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
