% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Solution to knapsack problem using Greedy Algorithm}
\usage{
greedy_knapsack(x, W, ...)
}
\arguments{
\item{x}{Data.frame Contains 2 columns "w" and "v" containing the weight of sacks and value of sacks respectively.}

\item{W}{Numeric. Specifies the capacity of knapsack.}

\item{...}{Optional logical parameter optimised with default=FALSE. If set to TRUE, an optimised version (developed using profvis analysis) of algorithm is used.}
}
\value{
Named List
}
\description{
The function implements solution to knapsack problem using Greedy Algorithm [wiki](https://en.wikipedia.org/wiki/Knapsack).
}
\examples{
#Prepare Data:
suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
knapsack_objects<-data.frame(w=sample(1:4000,size=1000000,replace=TRUE),v=runif(n=1000000,0,10000))

ptm <- proc.time()
a<-greedy_knapsack(x = knapsack_objects[1:1000000,], W = 2000)
print(proc.time() - ptm) # 5.807 milliseconds
print(a)

ptm <- proc.time()
b<-greedy_knapsack(x = knapsack_objects[1:1000000,], W = 2000, optimised=TRUE)
print(proc.time() - ptm) # 0.227 milliseconds
print(b)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
