% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Solution to knapsack problem using dynamic programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Data.frame Contains 2 columns "w" and "v" containing the weight of sacks and value of sacks respectively.}

\item{W}{Numeric. Specifies the capacity of knapsack.}
}
\value{
Named List
}
\description{
The function implements solution to knapsack problem using dynamic programming. [wiki](https://en.wikipedia.org/wiki/Knapsack).
}
\examples{
#Prepare Data:
suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
knapsack_objects<-data.frame(w=sample(1:4000,size=1000,replace=TRUE),v=runif(n=1000,0,10000))

ptm <- proc.time()
a<-knapsack_dynamic(x = knapsack_objects[1:500,], W = 2000)
print(proc.time() - ptm) # 1.808 milliseconds
print(a)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
